+++ docs/changelog.md
@@ -31,6 +31,7 @@ Click on the version in the following list to see all changes since the previous
 * refactor: rename estimate service in preparation for new custom field logic
 * refactor: rename story point field in preparation for new custom field logic
 * feat: calculation of bounded votes supports non-numeric card values (#102)
+* feat: simplify logic to calculate agreement reached flag (#102)
 
 ### [4.10.0] - 2020-03-28